Never permit the release of a PRT
The PRT will only ever be returned to the client
as an encrypted blob. This encrypted blob can be
imported and decrypted later, but is otherwise a
black box to the client. This decreases the risk
that the PRT could be leaked. Even if the PRT
were now stored in an unsafe location (such as in
a world readable database), it would be useless
without the TPM secrets.

Signed-off-by: David Mulder <dmulder@samba.org>