Make cbindgen generate ifdefs from cfgs
By default, cbindgen ignores conditional checks via the cfg attribute,
and includes all code which cfg attributes are attached to
unconditionally.

That's causing issues with the conditional parameters which are defined
in the C API functions on the condition that the "on_behalf_of" feature
is activated. cbindgen always includes these in the generated header,
which means the header is incompatible with code that doesn't expect
these API functions to take these parameters.

The libhimmelblau example project msal_example.c is affected by that
(and msal_example.py probably too, because I assume the Python bindings
are also generated from the C header, but I didn't check).

cbindgen supports the generation of ifdefs from cfg attributes by
specifying those in the [defines] section of the cbindgen.toml.

Unfortunately, there is an issue in cbindgen that it doesn't generate
these ifdefs for conditional parameters: https://github.com/mozilla/cbindgen/issues/1037

We should still specify the mappings from cfgs used in capi.rs to
defines in the [defines] section.