Fix build failures of C example project
cbindgen unconditionally includes the on_behalf_of parameter in the
generated header, even though we now specify the mapping from cfgs to
defines in the cbindgen.toml (see https://github.com/mozilla/cbindgen/issues/1037).

So to avoid build failures, we need to pass this parameter.

Note that this also means that the C example must be run with a
libhimmelblau.so which was built with the on_behalf_of feature.