Try reading error message from strServiceExceptionMessage
I encountered a case where the config returned in the body of a response
from the authorize endpoint did not contain a "sErrTxt" field, but it
did contain an error message in the "strServiceExceptionMessage" field.

It also was a case where the AADSTS error code was not in the list of
known codes which we map to error message (see aadsts_err_gen.rs), which
makes it even more useful to use the error message from the other field.

Before:

    AADSTSError(AADSTSError { code: 500113, description: "Unknown error code" })

After:

    AADSTSError(AADSTSError { code: 500113, description: "AADSTS500113: No reply address is registered for the application." })