MS requires we specify the new Intune protocol version
MS has always required that we specify a protocol
version, but a recent update caused them to
reject the version we were using. This patch
kicks that can down the road another year, while
also adding the ability to detect newer versions.
In Himmelblau, we'll need to make this
configurable, since it appears only certain
version strings are permitted (the latest version
works, but the oldest support version does not).

Signed-off-by: David Mulder <dmulder@samba.org>